requirements.txt
Flask
gunicorn
raven[flask]
pillow
google-cloud-storage
google-cloud-datastore
google-cloud-pubsub
pytest>=3